\name{andy}
\alias{andy}
\docType{data}
\title{
Big Andy's Burger Barn
}
\description{
obs:  75   a sample of hamburger franchises in 75 cities from Big Andy's Burger Barn.
}
\usage{data("andy")}
\format{
  A data frame with 75 observations on the following 3 variables.
  \describe{
    \item{\code{sales}}{Monthly sales revenue ($1000s)
}
    \item{\code{price}}{A price index for all products sold in a given month.
}
 \item{\code{advert}}{Expenditure on advertising ($1000s)

}
  }
}
\details{
N/A
}
\source{
Courtesy of hamburger franchise, called Big Andy's Burger Barn

}
\references{
N/A
}
\examples{
data(andy)
## maybe str(andy) ; plot(andy) ...
}
\keyword{datasets}
