\name{bachelor}
\alias{bachelor}
\docType{data}
\title{
GDP and bachelor degree at state level
}
\description{
State level measures from 2008, Census Bureau
}
\usage{data("bachelor")}
\format{
  A data frame with 51 observations on the following 3 variables.
  \describe{
    \item{\code{state}}{state}
    \item{\code{income}}{state per capita income, $2008}
    \item{\code{bachelor}}{percent of state population with bachelor }
  }
}
\details{
N/A
}
\source{
N/A
}
\references{
N/A
}
\examples{
data(bachelor)
## maybe str(bachelor) ; plot(bachelor) ...
}
\keyword{datasets}
