\name{bangla}
\alias{bangla}
\docType{data}
\title{
Sugar cane market in Bangladesh
}
\description{
Prices and areas cultivated, annual data.
}
\usage{data("bangla")}
\format{
  A data frame with 34 observations on the following 2 variables.
  \describe{
    \item{\code{p}}{Price of sugar cane divided by the price of jute (jute is an alternative crop to sugar cane, planted by Bangladesh farmers)
}
    \item{\code{a}}{Area of sugar cane planted in thousands of hectares in a region of Bangladesh}
  }
}
\details{
N/A
}
\source{
M.Ec. dissertation by M.Jafforulah at the University of New England
}
\references{
N/A
}
\examples{
data(bangla)
## maybe str(bangla) ; plot(bangla) ...
}
\keyword{datasets}
