\name{bangla5}
\alias{bangla5}
\docType{data}
\title{
Sugar cane market in Bangladesh (updated)
}
\description{
Prices and areas cultivated, annual data.
}
\usage{data("bangla5")}
\format{
  A data frame with 73 observations on the following 2 variables.
  \describe{
    \item{\code{area}}{Area of sugar cane planted in thousands of hectares in a region of Bangladesh}
    \item{\code{price}}{Price of sugar cane divided by the price of jute (jute is an alternative crop to sugar cane, planted by Bangladesh farmers)
}
  }
}
\details{
N/A
}
\source{
An update of data used in the M.Ec. dissertation by M.Jafforulah at the University of New England
}
\references{
N/A
}
\examples{
data(bangla5)
## maybe str(bangla5) ; plot(bangla5) ...
}
\keyword{datasets}
