\name{beer}
\alias{beer}
\docType{data}
\title{
Individual demand for beer
}
\description{Annual observations from a single household.
}
\usage{data("beer")}
\format{
  A data frame with 30 observations on the following 5 variables.
  \describe{
    \item{\code{q}}{litres of beer consumed}
    \item{\code{pb}}{Price of beer ($)}
    \item{\code{pl}}{price of other liquor ($)}
    \item{\code{pr}}{price of remaining goods and services (an index)
}
    \item{\code{i}}{income ($)}
  }
}
\details{
N/A
}
\source{
N/A
}
\references{
N/A
}
\examples{
data(beer)
## maybe str(beer) ; plot(beer) ...
}
\keyword{datasets}
