\name{byd}
\alias{byd}
\docType{data}
\title{
BrightenYourDay (BYD) Lighting
}
\description{
Generated data on returns to shares in BrightenYourDay (BYD) Lighting.
}
\usage{data("byd")}
\format{
  A data frame with 500 observations on the following variable.
  \describe{
    \item{\code{r}}{returns to shares in BrightenYourDay (BYD) Lighting}
  }
}
\details{
N/A
}
\source{
Generated data
}
\references{
N/A
}
\examples{
data(byd)
## maybe str(byd) ; plot(byd) ...
}
\keyword{datasets}
