\name{canada6}
\alias{canada6}
\docType{data}
\title{
USD - CAD exchange rate
}
\description{
Canada / U.S. Foreign Exchange Rate, monthly, (1971M1 - 2017M3).
}
\usage{data("canada6")}
\format{
  A data frame with 555 observations on the following 2 variables.
  \describe{
    \item{\code{dateid01}}{date 01monthyear}
    \item{\code{canada}}{Canada / U.S. Foreign Exchange Rate}
  }
}
\details{
N/A
}
\source{
Board of Governors of the Federal Reserve System
}
\references{
N/A
}
\examples{
data(canada6)
## maybe str(canada6) ; plot(canada6) ...
}
\keyword{U.S.; Canada; exchange rate}
