\name{commute5}
\alias{commute5}
\docType{data}
\title{
Commuting times
}
\description{
Daily observations for each working day in 2015.
}
\usage{data("commute5")}
\format{
  A data frame with 249 observations on the following 4 variables.
  \describe{
    \item{\code{time}}{commuting time in minutes}
    \item{\code{depart}}{number of minutes after 6.30AM that Bill  leaves for work}
    \item{\code{reds}}{number of red traffic lights that Bill encounters}
    \item{\code{trains}}{number of trains that Bill has to wait for at Murrumbeena railway station}
  }
}
\details{
N/A
}
\source{
Commuting time records kept by Professor B. Griffiths

}
\references{
N/A
}
\examples{
data(commute5)
## maybe str(commute5) ; plot(commute5) ...
}
\keyword{commuting time}
