\name{csi}
\alias{csi}
\docType{data}
\title{
Consymer sentiment index
}
\description{
Monthly, (1978:01 - 2006:12).
}
\usage{data("csi")}
\format{
  A data frame with 348 observations on the following variable.
  \describe{
    \item{\code{csi}}{Consumer Sentiment Index}
  }
}
\details{
N/A
}
\source{
Survey Research Center, University of Michigan
}
\references{
N/A
}
\examples{
data(csi)
## maybe str(csi) ; plot(csi) ...
}
\keyword{consumer sentiment index}
