\name{equity}
\alias{equity}
\docType{data}
\title{
S&P Composite Index
}
\description{
Annual data, (1889 - 1979).
}
\usage{data("equity")}
\format{
  A data frame with 91 observations on the following 2 variables.
  \describe{
    \item{\code{dn}}{nominal dividends per share}
    \item{\code{pn}}{Standard and Poor Composite Price Index}
  }
}
\details{
N/A
}
\source{
Prescott,E.C. and Mehra,R. "The Equity Premium: A Puzzle," Journal of Monetary Economics, 15 March, 1985, pp. 145-161.
}
\references{
N/A
}
\examples{
data(equity)
## maybe str(equity) ; plot(equity) ...
}
\keyword{S&P composite index}
