\name{euro}
\alias{euro}
\docType{data}
\title{
Euro Share price index
}
\description{
monthly (1988:01-2004:12)
}
\usage{data("euro")}
\format{
  A data frame with 204 observations on the following variable.
  \describe{
    \item{\code{r}}{returns to the Euro Share price Index}
  }
}
\details{
N/A
}
\source{
http://au.finance.yahoo.com
}
\references{
N/A
}
\examples{
data(euro)
## maybe str(euro) ; plot(euro) ...
}
\keyword{euro share}
