\name{exrate5}
\alias{exrate5}
\docType{data}
\title{
USD - AUD exchange rate
}
\description{
Monthly, (1985:07 - 2010:06).
}
\usage{data("exrate5")}
\format{
  A data frame with 300 observations on the following 2 variables.
  \describe{
    \item{\code{dateid01}}{date}
    \item{\code{s}}{change in the $US/$A exchange rate}
  }
}
\details{
N/A
}
\source{
Reserve Bank of Australia, Statistical Bulletin
}
\references{
N/A
}
\examples{
data(exrate5)
## maybe str(exrate5) ; plot(exrate5) ...
}
\keyword{australia exchange rate}
