\name{food}
\alias{food}
\docType{data}
\title{
Food expenditure
}
\description{
Food expenditure and income,individual data
}
\usage{data("food")}
\format{
  A data frame with 40 observations on the following 2 variables.
  \describe{
    \item{\code{food_exp}}{weekly food expenditure in $
}
    \item{\code{income}}{weekly income in $100}
  }
}
\details{
N/A
}
\source{
N/A
}
\references{
N/A
}
\examples{
data(food)
## maybe str(food) ; plot(food) ...
}
\keyword{food expenditure}
