\name{gdp}
\alias{gdp}
\docType{data}
\title{
U.S. and Australian GDP
}
\description{
Quarterly, (1970:1 - 2000.4).
}
\usage{data("gdp")}
\format{
  A data frame with 124 observations on the following 2 variables.
  \describe{
    \item{\code{usa}}{real GDP of USA}
    \item{\code{aus}}{real GDP of Australia}
  }
}
\details{
N/A
}
\source{
International Financial Statistics, International Monetary Fund

}
\references{
N/A
}
\examples{
data(gdp)
## maybe str(gdp) ; plot(gdp) ...
}
\keyword{gdp U.S. Australia}
