\name{gdp5}
\alias{gdp5}
\docType{data}
\title{
U.S. GDP
}
\description{
quarterly observations, U.S. data from 1984Q1 to 2016Q4
}
\usage{data("gdp5")}
\format{
  A data frame with 132 observations on the following 2 variables.
  \describe{
    \item{\code{dateid01}}{date}
    \item{\code{gdp}}{real gross domestic product for the U.S. in trillions of 2009 dollars (FRED series GDPC1)
}
  }
}
\details{
N/A
}
\source{
Federal Reserve Bank of St Louis (FRED)
}
\references{
N/A
}
\examples{
data(gdp5)
## maybe str(gdp5) ; plot(gdp5) ...
}
\keyword{datasets}
