\name{gfc}
\alias{gfc}
\docType{data}
\title{
U.S. and Euro Area GDP
}
\description{
quarterly (1995q1 - 2009q4)
}
\usage{data("gfc")}
\format{
  A data frame with 60 observations on the following 2 variables.
  \describe{
    \item{\code{leuro}}{log of gross domestic product for the Euro Area}
    \item{\code{lusa}}{log of gross domestic product for the United States
}
  }
}
\details{
N/A
}
\source{
Statistical Bulletin, Reserve Bank of Australia
}
\references{
N/A
}
\examples{
data(gfc)
## maybe str(gfc) ; plot(gfc) ...
}
\keyword{datasets}
