\name{gold}
\alias{gold}
\docType{data}
\title{
Gold company shares
}
\description{
Daily, Dec 13, 2005 to Sept 18, 2006.
}
\usage{data("gold")}
\format{
  A data frame with 200 observations on the following variable.
  \describe{
    \item{\code{gold}}{returns to shares to a company specialising in gold bullions   }
  }
}

\source{
http://au.finance.yahoo.com
}

\examples{
data(gold)
## maybe str(gold) ; plot(gold) ...
}
\keyword{gold}
