\name{insure}
\alias{insure}
\docType{data}
\title{
Insurance and income
}
\description{
Twenty hypothetical households
}
\usage{data("insure")}
\format{
  A data frame with 20 observations on the following 2 variables.
  \describe{
    \item{\code{insurance}}{thousands of dollars }
    \item{\code{income}}{thousands of dollars}
  }
}

\examples{
data(insure)
## maybe str(insure) ; plot(insure) ...
}
\keyword{insurance}
