\name{inter2}
\alias{inter2}
\docType{data}
\title{
Generated time series
}
\description{
Generated time series, 300 observations.
}
\usage{data("inter2")}
\format{
  A data frame with 300 observations on the following variable.
  \describe{
    \item{\code{y}}{a numeric vector}
  }
}

\examples{
data(inter2)
## maybe str(inter2) ; plot(inter2) ...
}
\keyword{datasets}
