\name{liquor5}
\alias{liquor5}
\docType{data}
\title{
Liquor Consumption
}
\description{
Expenditure on liquor by household.
}
\usage{data("liquor5")}
\format{
  A data frame with 120 observations on the following 4 variables.
  \describe{
    \item{\code{hh}}{household}
    \item{\code{year}}{a numeric vector}
    \item{\code{liquor}}{liquor, thousands of $}
    \item{\code{income}}{income, thousands of $}
  }
}
\examples{
data(liquor5)
## maybe str(liquor5) ; plot(liquor5) ...
}
\keyword{datasets}
