\name{mc100}
\alias{mc100}
\docType{data}
\title{
Simulated data
}
\description{Simulated, 100 observations.}
\usage{data("mc100")}
\format{
  A data frame with 100 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}

\examples{
data(mc100)
## maybe str(mc100) ; plot(mc100) ...
}
\keyword{datasets}
