\name{mc1000}
\alias{mc1000}
\docType{data}
\title{
Simulated data
}
\description{Simulated, 1000 observations of two variables.}
\usage{data("mc1000")}
\format{
  A data frame with 1000 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}

\examples{
data(mc1000)
## maybe str(mc1000) ; plot(mc1000) ...
}
\keyword{datasets}
