\name{mc20}
\alias{mc20}
\docType{data}
\title{
Simulated data
}
\description{
Simulated, 20 observations.
}
\usage{data("mc20")}
\format{
  A data frame with 20 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}

\examples{
data(mc20)
## maybe str(mc20) ; plot(mc20) ...
}
\keyword{datasets}
