\name{mc200}
\alias{mc200}
\docType{data}
\title{
Simulated data
}
\description{Simulated, 200 observations}
\usage{data("mc200")}
\format{
  A data frame with 200 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}

\examples{
data(mc200)
## maybe str(mc200) ; plot(mc200) ...
}
\keyword{datasets}
