\name{mc40}
\alias{mc40}
\docType{data}
\title{Simulated data}
\description{Simulated, 40 observations.}
\usage{data("mc40")}
\format{
  A data frame with 40 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}

\examples{
data(mc40)
## maybe str(mc40) ; plot(mc40) ...
}
\keyword{datasets}
