\name{mc500}
\alias{mc500}
\docType{data}
\title{
Simulated data
}
\description{Simulated, 500 observations.}
\usage{data("mc500")}
\format{
  A data frame with 500 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}

\examples{
data(mc500)
## maybe str(mc500) ; plot(mc500) ...
}
\keyword{datasets}
