\name{mexico}
\alias{mexico}
\docType{data}
\title{GDPs of Mexico and U.S.}
\description{Quarterly, (1980:1 - 2006.3.)  }
\usage{data("mexico")}
\format{
  A data frame with 107 observations on the following 2 variables.
  \describe{
    \item{\code{mexico}}{real GDP of Mexico   (standardised value of 100 in 2000)
}
    \item{\code{usa}}{real GDP of the U.S. (standardised value of 100 in 2000)
}
  }
}

\examples{
data(mexico)
## maybe str(mexico) ; plot(mexico) ...
}
\keyword{datasets}
