\name{nlls}
\alias{nlls}
\docType{data}
\title{
NLS data for Example 6.19
}
\description{
These are three variables used in the textbook example 6.19 to demonstrate the use of nonlinear least squares method.
}
\usage{data("nlls")}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{y}}{a numeric vector}
    \item{\code{x1}}{a numeric vector}
    \item{\code{x2}}{a numeric vector}
  }
}
\examples{
data(nlls)
## maybe str(nlls) ; plot(nlls) ...
}
\keyword{NLS}
