\name{share}
\alias{share}
\docType{data}
\title{Singapore}
\description{204, monthly (1988:01-2004:12)  }
\usage{data("share")}
\format{
  A data frame with 204 observations on the following variable.
  \describe{
    \item{\code{y}}{Straits Times Share Price Index of Singapore}
  }
}

\source{Bloomberg}

\examples{
data(share)
## maybe str(share) ; plot(share) ...
}
\keyword{datasets}
