\name{sp}
\alias{sp}
\docType{data}
\title{S&P500
}
\description{781, weekly, (1.08.1990 - 12.20.2004)
}
\usage{data("sp")}
\format{
  A data frame with 781 observations on the following variable.
  \describe{
    \item{\code{r}}{returns to US S&P500 }
  }
}

\source{http://au.finance.yahoo.com  }

\examples{
data(sp)
## maybe str(sp) ; plot(sp) ...
}
\keyword{datasets}
