\name{spurious}
\alias{spurious}
\docType{data}
\title{Random walk}
\description{Two generated series}
\usage{data("spurious")}
\format{
  A data frame with 700 observations on the following 2 variables.
  \describe{
    \item{\code{rw1}}{random walk process}
    \item{\code{rw2}}{random walk process}
  }
}

\examples{
data(spurious)
## maybe str(spurious) ; plot(spurious) ...
}
\keyword{datasets}
