\name{table2_2}
\alias{table2_2}
\docType{data}
\title{Food expenditure, 10 households}
\description{This is a subset of the 'food' dataset}
\usage{data("table2_2")}
\format{
  A data frame with 40 observations on the following 11 variables.
  \describe{
    \item{\code{x}}{weekly income in $100}
    \item{\code{y1}}{y1,...,y10 are food expnditure for 10 households}
    \item{\code{y2}}{a numeric vector}
    \item{\code{y3}}{a numeric vector}
    \item{\code{y4}}{a numeric vector}
    \item{\code{y5}}{a numeric vector}
    \item{\code{y6}}{a numeric vector}
    \item{\code{y7}}{a numeric vector}
    \item{\code{y8}}{a numeric vector}
    \item{\code{y9}}{a numeric vector}
    \item{\code{y10}}{a numeric vector}
  }
}

\examples{
data(table2_2)
## maybe str(table2_2) ; plot(table2_2) ...
}
\keyword{datasets}
