\name{uk}
\alias{uk}
\docType{data}
\title{CPI for UK}
\description{589, monthly (1957:06-2006:06)  }
\usage{data("uk")}
\format{
  A data frame with 589 observations on the following variable.
  \describe{
    \item{\code{ukcpi}}{Consumer Price Index of the United Kingdom (seasonally adjusted)
}
  }
}

\source{International Financial Statistics, IMF}

\examples{
data(uk)
## maybe str(uk) ; plot(uk) ...
}
\keyword{datasets}
