\name{unit}
\alias{unit}
\docType{data}
\title{Random processes}
\description{199 observations}
\usage{data("unit")}
\format{
  A data frame with 199 observations on the following 4 variables.
  \describe{
    \item{\code{w}}{AR process}
    \item{\code{x}}{AR process with trend}
    \item{\code{y}}{Random walk}
    \item{\code{z}}{Random walk with drift}
  }
}

\source{generated series }

\examples{
data(unit)
## maybe str(unit) ; plot(unit) ...
}
\keyword{datasets}
