\name{vacation}
\alias{vacation}
\docType{data}
\title{Vacation travellers}
\description{A sample of Chicago households}
\usage{data("vacation")}
\format{
  A data frame with 200 observations on the following 4 variables.
  \describe{
    \item{\code{miles}}{miles traveled per year}
    \item{\code{income}}{annual income in $1000's}
    \item{\code{age}}{average age of adult members of household
}
    \item{\code{kids}}{number of children in household}
  }
}

\examples{
data(vacation)
## maybe str(vacation) ; plot(vacation) ...
}
\keyword{datasets}
