\name{var}
\alias{var}
\docType{data}
\title{Two random walk processes}
\description{100 generated observations for two variables}
\usage{data("var")}
\format{
  A data frame with 100 observations on the following 2 variables.
  \describe{
    \item{\code{w}}{random walk process}
    \item{\code{z}}{random walk process}
  }
}
\details{
N/A
}
\source{generated time series data}
\references{
N/A
}
\examples{
data(var)
## maybe str(var) ; plot(var) ...
}
\keyword{datasets}
