\name{vec}
\alias{vec}
\docType{data}
\title{Two random walk processes}
\description{100 generated observations for two variables}
\usage{data("vec")}
\format{
  A data frame with 100 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{random walk process}
    \item{\code{y}}{random walk process}
  }
}

\source{generated time series data}

\examples{
data(vec)
## maybe str(vec) ; plot(vec) ...
}
\keyword{datasets}
