\name{wa_wheat}
\alias{wa_wheat}
\docType{data}
\title{Wheat yield in Australia}
\description{48 observations}
\usage{data("wa_wheat")}
\format{
  A data frame with 48 observations on the following 5 variables.
  \describe{
    \item{\code{northampton}}{average wheat yield in tonnes per hectare, Northampton Shire, Western Australia}
    \item{\code{chapman}}{average wheat yield in tonnes per hectare, Chapman Shire, Western Australia}
    \item{\code{mullewa}}{average wheat yield in tonnes per hectare, Mullewa Shire, Western Australia}
    \item{\code{greenough}}{average wheat yield in tonnes per hectare, Greenough Shire, Western Australia}
    \item{\code{time}}{time index, 1,...,48}
  }
}

\examples{
data(wa_wheat)
## maybe str(wa_wheat) ; plot(wa_wheat) ...
}
\keyword{datasets}
